Cleaned utils (or did I)

+++ odinclient/src/main/kotlin/me/odinclient/features/impl/floor7/p3/SimonSays.kt
@@ -120,7 +120,7 @@ object SimonSays : Module(
     fun onEntityJoin(event: PostEntityMetadata) {
         val ent = mc.theWorld.getEntityByID(event.packet.entityId)
         if (ent !is EntityItem || Item.getIdFromItem(ent.entityItem.item) != 77) return
-        val pos = BlockPos(ent.posX.floor(), ent.posY.floor(), ent.posZ.floor()).east()
+        val pos = BlockPos(ent.posX.floor().toDouble(), ent.posY.floor().toDouble(), ent.posZ.floor().toDouble()).east()
         val index = clickInOrder.indexOf(pos)
         if (index == 2 && clickInOrder.size == 3) {
             clickInOrder.removeFirst()
