+++ odinmain/src/main/kotlin/me/odinmain/features/impl/render/BPSDisplay.kt
@@ -44,7 +44,7 @@ object BPSDisplay : Module(
     fun tick(event: ClientTickEvent) {
         if (!isBreaking) return
         val secondsElapsed = (System.currentTimeMillis() - startTime) / 1000.0
-        bps = (blocksBroken / secondsElapsed).round(2)
+        bps = (blocksBroken / secondsElapsed).round(2).toDouble()
         if (System.currentTimeMillis() - lastBrokenBlock > 1000) {
             bps = 0.0
             isBreaking = false
